Add files via upload
This code is a demonstration of deep learning based on a single neural node. Featured in the files is Kacmarcz's algorithm (first-order gradient descent) and DFP which was first developed in 1959 and then further innovated in 1963. The DFP algorithm works by approximating the Hessian matrix (a second-order gradient descent). In this manner, it can converge to optimal weights in a single episode.